+++ frontend/src/hooks/use-workflow-tree.ts
@@ -1,0 +1,125 @@
+import { useCallback, useState } from "react";
+import type { FeatureType, WorkflowNode, WorkflowTree } from "@/types/research";
+
+const initialWorkflowTree: WorkflowTree = {
+  nodes: {},
+  rootId: null,
+  activeNodeId: null,
+};
+
+export function useWorkflowTree() {
+  const [workflowTree, setWorkflowTree] = useState<WorkflowTree>(initialWorkflowTree);
+  const [activeNodeId, setActiveNodeId] = useState<string | null>(null);
+
+  const handleNavigate = useCallback(
+    (nodeId: string) => {
+      const node = workflowTree.nodes[nodeId];
+      if (node) {
+        setActiveNodeId(nodeId);
+      }
+    },
+    [workflowTree],
+  );
+
+  const addWorkflowNode = useCallback(
+    (
+      type: FeatureType,
+      parentNodeId: string | null,
+      isBranch = false,
+      data?: WorkflowNode["data"],
+      snapshot?: WorkflowNode["snapshot"],
+    ): string => {
+      const newNodeId = `node-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
+
+      setWorkflowTree((prev) => {
+        const newNodes = { ...prev.nodes };
+
+        let branchIndex = 0;
+        if (parentNodeId) {
+          const parentNode = prev.nodes[parentNodeId];
+          if (parentNode) {
+            if (isBranch) {
+              const siblingBranches = parentNode.children.map(
+                (id) => prev.nodes[id]?.branchIndex || 0,
+              );
+              const maxSiblingBranch =
+                siblingBranches.length > 0 ? Math.max(...siblingBranches) : parentNode.branchIndex;
+              branchIndex = maxSiblingBranch + 1;
+            } else {
+              branchIndex = parentNode.branchIndex;
+            }
+          }
+        } else if (isBranch) {
+          let maxBranch = 0;
+          for (const nodeId in prev.nodes) {
+            maxBranch = Math.max(maxBranch, prev.nodes[nodeId].branchIndex);
+          }
+          branchIndex = maxBranch + 1;
+        }
+
+        const newNode: WorkflowNode = {
+          id: newNodeId,
+          type,
+          branchIndex,
+          parentId: parentNodeId,
+          children: [],
+          data: data || {},
+          snapshot,
+          createdAt: new Date(),
+        };
+
+        newNodes[newNodeId] = newNode;
+
+        if (parentNodeId && newNodes[parentNodeId]) {
+          newNodes[parentNodeId] = {
+            ...newNodes[parentNodeId],
+            children: [...newNodes[parentNodeId].children, newNodeId],
+          };
+        }
+
+        return {
+          nodes: newNodes,
+          rootId: prev.rootId || newNodeId,
+          activeNodeId: newNodeId,
+        };
+      });
+
+      setActiveNodeId(newNodeId);
+      return newNodeId;
+    },
+    [],
+  );
+
+  const updateNodeSnapshot = useCallback((nodeId: string, snapshot: WorkflowNode["snapshot"]) => {
+    setWorkflowTree((prev) => ({
+      ...prev,
+      nodes: {
+        ...prev.nodes,
+        [nodeId]: {
+          ...prev.nodes[nodeId],
+          snapshot,
+        },
+      },
+    }));
+  }, []);
+
+  const resetDownstreamSessions = useCallback((_fromType: FeatureType) => {
+    // ここはそのまま
+  }, []);
+
+  const resetWorkflow = useCallback(() => {
+    setWorkflowTree(initialWorkflowTree);
+    setActiveNodeId(null);
+  }, []);
+
+  return {
+    workflowTree,
+    activeNodeId,
+    setActiveNodeId,
+    handleNavigate,
+    addWorkflowNode,
+    updateNodeSnapshot,
+    resetDownstreamSessions,
+    resetWorkflow,
+  };
+}
