+++ frontend/src/main.tsx
@@ -8,6 +8,20 @@ import { OpenAPI } from "./lib/api";
 
 OpenAPI.BASE = (import.meta.env.VITE_API_BASE_URL as string | undefined) ?? "http://localhost:8000";
 
+// Attach GitHub session header only to GitHub-related generated API calls
+OpenAPI.HEADERS = async (options) => {
+  const sessionToken = localStorage.getItem("github_session_token");
+  const headers: Record<string, string> = {};
+  if (
+    sessionToken &&
+    typeof options?.url === "string" &&
+    options.url.toLowerCase().includes("github")
+  ) {
+    headers["X-GitHub-Session"] = sessionToken;
+  }
+  return headers;
+};
+
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
     <BrowserRouter>
